Document 'msg' param and fix examples
The 'msg' alias for 'subject' isn't in the documentation, so adding it.

In the gmail example, it uses both the 'subject' and 'msg' params, but 'msg' is an alias of 'subject', so you are essentially declaring the same param twice. If you use this example, then no subject is sent (I tested with gmail). Documentation example is updated to use 'body' as intended.

Also, updated the simple example to use 'subject' instead of the 'msg' alias since it is more explicit.